+++ core/user/db-adapter/index.ts
@@ -1,1 +1,3 @@
 export * from './user.entity';
+
+export * from './view/users-count.view-entity';