+++ core/user/domain/index.ts
@@ -5,3 +5,5 @@ export * from './utils/user.exceptions';
 export * from './utils/user.mapper';
 export * from './dto/user.dto';
 export * from './events';
+
+export * from './types/user-analytics.types';