+++ core/user/domain/types/user-analytics.types.ts
@@ -1,0 +1,1 @@
+export type IUsersCountViewEntity = { count: number };