+++ packages/ping/ping.controller.ts
@@ -2,9 +2,7 @@ import { Controller, Get, Logger } from '@nestjs/common';
 
 @Controller('/ping')
 export class PingController {
-  constructor(protected readonly logger: Logger) {
-    this.logger.error('test', 'test');
-  }
+  constructor(protected readonly logger: Logger) {}
 
   @Get()
   public async get(): Promise<string> {