+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @muzikanto/backend-starter
 
+## 0.2.0
+
+### Minor Changes
+
+- 032f134: open telemetry module
+- 913f145: auth module added
+- 0f1329d: fix alertmanager
+
+### Patch Changes
+
+- add typeorm view entity example
+
 ## 0.1.0
 
 ### Minor Changes